# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: B4P Youth API. Member routes require an active signed-in youth account; /admin routes require an active administrator. About and contact submission are open.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /applications:
    get:
      operationId: listApplicationForms
      tags: [applications]
      summary: List published application forms
      parameters:
        - in: query
          name: programId
          schema: { type: integer }
      responses:
        "200":
          description: Published application forms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApplicationForm"
  /applications/{id}:
    get:
      operationId: getApplicationForm
      tags: [applications]
      summary: Get a published application form and its questions
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Application form
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationForm"
        "404": { $ref: "#/components/responses/NotFound" }
  /applications/{id}/submissions:
    post:
      operationId: submitApplicationForm
      tags: [applications]
      summary: Submit answers to a published application form
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationSubmissionInput"
      responses:
        "201":
          description: Submitted application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationSubmissionReceipt"
        "409": { $ref: "#/components/responses/Conflict" }
...
  /admin/application-forms:
    get:
      operationId: listAdminApplicationForms
      tags: [admin, applications]
      summary: List application forms, including drafts
      responses:
        "200":
          description: Application forms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApplicationForm"
    post:
      operationId: createApplicationForm
      tags: [admin, applications]
      summary: Create an application form with custom questions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationFormInput"
      responses:
        "201":
          description: Created application form
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationForm"
  /admin/application-forms/{id}:
    patch:
      operationId: updateApplicationForm
      tags: [admin, applications]
      summary: Edit, publish or unpublish an application form
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationFormUpdate"
      responses:
        "200":
          description: Updated application form
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationForm"
    delete:
      operationId: deleteApplicationForm
      tags: [admin, applications]
      summary: Delete an application form while retaining submitted records
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204": { description: Application form deleted }
  /admin/application-submissions:
    get:
      operationId: listAdminApplicationSubmissions
      tags: [admin, applications]
      summary: Search and filter application form submissions
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: status
          schema:
            $ref: "#/components/schemas/ApplicationStatus"
        - in: query
          name: formId
          schema: { type: integer }
      responses:
        "200":
          description: Submissions with applicant details and saved answers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminApplicationSubmission"
  /admin/application-submissions/{id}:
    patch:
      operationId: updateApplicationSubmission
      tags: [admin, applications]
      summary: Update a submission status and internal note
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationSubmissionUpdate"
      responses:
        "200":
          description: Updated application submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminApplicationSubmission"
...
components:
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema: { type: integer, minimum: 1 }
  responses:
    NotFound:
      description: Record not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: Duplicate or conflicting request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    ApplicationQuestionType:
      type: string
      enum: [short_text, long_text, email, phone, dropdown, multiple_choice, checkbox, date]
    ApplicationStatus:
      type: string
      enum: [Submitted, Under Review, Accepted, Rejected]
...
    ApplicationQuestion:
      type: object
      required: [id, prompt, type, required, options, position]
      properties:
        id: { type: integer }
        prompt: { type: string }
        type: { $ref: "#/components/schemas/ApplicationQuestionType" }
        required: { type: boolean }
        options:
          type: array
          items: { type: string }
        position: { type: integer }
    ApplicationQuestionInput:
      type: object
      required: [prompt, type, required, options, position]
      properties:
        prompt: { type: string, minLength: 2, maxLength: 300 }
        type: { $ref: "#/components/schemas/ApplicationQuestionType" }
        required: { type: boolean }
        options:
          type: array
          items: { type: string, minLength: 1, maxLength: 200 }
        position: { type: integer, minimum: 0 }
    ApplicationForm:
      type: object
      required: [id, title, description, instructions, eligibilityRequirements, deadline, programId, published, createdAt, updatedAt, questions]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string }
        instructions: { type: string }
        eligibilityRequirements: { type: string }
        deadline: { type: ["string", "null"], format: date }
        programId: { type: ["integer", "null"] }
        published: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        questions:
          type: array
          items: { $ref: "#/components/schemas/ApplicationQuestion" }
    ApplicationFormInput:
      type: object
      required: [title, description, instructions, eligibilityRequirements, deadline, programId, published, questions]
      properties:
        title: { type: string, minLength: 2, maxLength: 200 }
        description: { type: string, minLength: 1, maxLength: 5000 }
        instructions: { type: string, maxLength: 5000 }
        eligibilityRequirements: { type: string, maxLength: 5000 }
        deadline: { type: ["string", "null"], format: date }
        programId: { type: ["integer", "null"] }
        published: { type: boolean }
        questions:
          type: array
          items: { $ref: "#/components/schemas/ApplicationQuestionInput" }
    ApplicationFormUpdate:
      type: object
      properties:
        title: { type: string, minLength: 2, maxLength: 200 }
        description: { type: string, minLength: 1, maxLength: 5000 }
        instructions: { type: string, maxLength: 5000 }
        eligibilityRequirements: { type: string, maxLength: 5000 }
        deadline: { type: ["string", "null"], format: date }
        programId: { type: ["integer", "null"] }
        published: { type: boolean }
        questions:
          type: array
          items: { $ref: "#/components/schemas/ApplicationQuestionInput" }
    ApplicationSubmissionInput:
      type: object
      required: [answers]
      properties:
        answers:
          type: object
          additionalProperties: true
    ApplicationAnswer:
      type: object
      required: [prompt, type, value]
      properties:
        prompt: { type: string }
        type: { $ref: "#/components/schemas/ApplicationQuestionType" }
        value:
          oneOf:
            - type: string
            - type: array
              items: { type: string }
            - type: boolean
    MyApplicationSubmission:
      type: object
      required: [id, formId, formTitle, status, adminNote, submittedAt]
      properties:
        id: { type: integer }
        formId: { type: ["integer", "null"] }
        formTitle: { type: string }
        status: { $ref: "#/components/schemas/ApplicationStatus" }
        adminNote: { type: ["string", "null"] }
        submittedAt: { type: string, format: date-time }
    ApplicationSubmissionReceipt:
      allOf:
        - $ref: "#/components/schemas/MyApplicationSubmission"
    AdminApplicationSubmission:
      allOf:
        - $ref: "#/components/schemas/MyApplicationSubmission"
        - type: object
          required: [userId, applicantName, applicantEmail, applicantLocation, answers]
          properties:
            userId: { type: integer }
            applicantName: { type: string }
            applicantEmail: { type: string, format: email }
            applicantLocation: { type: string }
            answers:
              type: array
              items: { $ref: "#/components/schemas/ApplicationAnswer" }
    ApplicationSubmissionUpdate:
      type: object
      required: [status]
      properties:
        status: { $ref: "#/components/schemas/ApplicationStatus" }
        adminNote: { type: ["string", "null"], maxLength: 3000 }
...