# lib/api-zod/src/generated/types/adminApplicationSubmission.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * B4P Youth API. Member routes require an active signed-in youth account; /admin routes require an active administrator. About and contact submission are open.
 * OpenAPI spec version: 0.1.0
 */
import type { ApplicationAnswer } from './applicationAnswer';
import type { MyApplicationSubmission } from './myApplicationSubmission';

export type AdminApplicationSubmission = MyApplicationSubmission & {
  userId: number;
  applicantName: string;
  applicantEmail: string;
  applicantLocation: string;
  answers: ApplicationAnswer[];
};