# lib/api-zod/src/generated/types/applicationAnswer.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * B4P Youth API. Member routes require an active signed-in youth account; /admin routes require an active administrator. About and contact submission are open.
 * OpenAPI spec version: 0.1.0
 */
import type { ApplicationQuestionType } from './applicationQuestionType';

export interface ApplicationAnswer {
  prompt: string;
  type: ApplicationQuestionType;
  value: string | string[] | boolean;
}