# lib/api-zod/src/generated/types/applicationFormInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * B4P Youth API. Member routes require an active signed-in youth account; /admin routes require an active administrator. About and contact submission are open.
 * OpenAPI spec version: 0.1.0
 */
import type { ApplicationQuestionInput } from './applicationQuestionInput';

export interface ApplicationFormInput {
  /**
     * @minLength 2
     * @maxLength 200
     */
  title: string;
  /**
     * @minLength 1
     * @maxLength 5000
     */
  description: string;
  /** @maxLength 5000 */
  instructions: string;
  /** @maxLength 5000 */
  eligibilityRequirements: string;
  /** @nullable */
  deadline: Date | null;
  /** @nullable */
  programId: number | null;
  published: boolean;
  questions: ApplicationQuestionInput[];
}